moved location of new describe functions
