+++ pyaurorax/util/__init__.py
@@ -1,7 +1,7 @@
 """
 The util module provides helper methods such as converting
 arbitrary geographic locations to North/South B-trace geographic
-locations, and search query "SQL-like" descriptions.
+locations.
 
 Note that all functions and classes from submodules are all imported
 at this level of the util module. They can be referenced from
@@ -11,9 +11,6 @@ here instead of digging in deeper to the submodules.
 # function and class imports
 from .calculate_btrace import (ground_geo_to_nbtrace,
                                ground_geo_to_sbtrace)
-from .describe import (describe_conjunction_search,
-                       describe_data_products_search,
-                       describe_ephemeris_search)
 
 # pdoc imports and exports
 from .calculate_btrace import __pdoc__ as __btrace_pdoc__
@@ -21,7 +18,4 @@ __pdoc__ = __btrace_pdoc__
 __all__ = [
     "ground_geo_to_nbtrace",
     "ground_geo_to_sbtrace",
-    "describe_conjunction_search",
-    "describe_data_products_search",
-    "describe_ephemeris_search",
 ]
