+++ tests/test_suite/conjunctions/test_conjunctions.py
@@ -613,3 +613,38 @@ def test_conjunctions_search_asynchronous_with_epoch_precision(precision_value):
     # check to make sure we got at least one result, and the
     # first result is a Conjunction object
     assert len(s.data) > 0 and type(s.data[0]) is Conjunction
+
+
+@pytest.mark.conjunctions
+def test_conjunction_search_describe():
+    # set params
+    start = datetime.datetime(2020, 1, 1, 0, 0, 0)
+    end = datetime.datetime(2020, 1, 1, 23, 59, 59)
+    ground_params = [
+        {"programs": ["themis-asi"]},
+    ]
+    space_params = [
+        {"programs": ["swarm", "themis"]},
+    ]
+    distance = 500
+    expected_response_str = "Find conjunctions of type (nbtrace) with epoch precision " \
+        "of 60 seconds between data sources of ground1=(program in (themis-asi)) AND " \
+        "space1=(program in (swarm, themis)) WHERE epochs are between 2020-01-01T00:00:00 " \
+        "AND 2020-01-01T23:59:59 UTC HAVING max distances between location points of " \
+        "ground1-space1=500 km."
+
+    # create search object
+    s = pyaurorax.conjunctions.Search(start,
+                                      end,
+                                      distance,
+                                      ground=ground_params,
+                                      space=space_params)
+
+    # get describe string
+    describe_str = pyaurorax.conjunctions.describe(s)
+
+    # test response
+    if (describe_str is not None and describe_str == expected_response_str):
+        assert True
+    else:
+        assert False
